CU-868evjqkw: MXRAndroidSystem Partial Classes
