+++ Assets/MXR.SDK/Runtime/Android/MXRAndroidSystem.Commands.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 86e1ac62a3b4c4a9d983bb108a1b69cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
