+++ Assets/MXR.SDK/Runtime/Android/MXRAndroidSystem.Messages.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3696ddb163a494070b0474f49b61f15a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
